@@ -33,27 +33,27 @@
   },
   "devDependencies": {
     "@angular-builders/custom-webpack": "^13.0.0",
-    "@angular-devkit/build-angular": "^13.0.2",
-    "@angular-devkit/core": "^13.0.2",
-    "@angular-devkit/schematics": "^13.0.2",
+    "@angular-devkit/build-angular": "^14.0.3",
+    "@angular-devkit/core": "^14.0.3",
+    "@angular-devkit/schematics": "^14.0.3",
     "@angular-eslint/builder": "^13.0.1",
     "@angular-eslint/eslint-plugin": "^13.0.1",
     "@angular-eslint/eslint-plugin-template": "^13.0.1",
     "@angular-eslint/schematics": "^13.0.1",
     "@angular-eslint/template-parser": "^13.0.1",
-    "@angular/animations": "^13.0.1",
-    "@angular/cli": "^13.0.2",
-    "@angular/common": "^13.0.1",
-    "@angular/compiler": "^13.0.1",
-    "@angular/compiler-cli": "^13.0.1",
-    "@angular/core": "^13.0.1",
-    "@angular/forms": "^13.0.1",
-    "@angular/language-service": "^13.0.1",
-    "@angular/platform-browser": "^13.0.1",
-    "@angular/platform-browser-dynamic": "^13.0.1",
-    "@angular/platform-server": "^13.0.1",
-    "@angular/router": "^13.0.1",
-    "@angular/service-worker": "^13.0.1",
+    "@angular/animations": "^14.0.3",
+    "@angular/cli": "^14.0.3",
+    "@angular/common": "^14.0.3",
+    "@angular/compiler": "^14.0.3",
+    "@angular/compiler-cli": "^14.0.3",
+    "@angular/core": "^14.0.3",
+    "@angular/forms": "^14.0.3",
+    "@angular/language-service": "^14.0.3",
+    "@angular/platform-browser": "^14.0.3",
+    "@angular/platform-browser-dynamic": "^14.0.3",
+    "@angular/platform-server": "^14.0.3",
+    "@angular/router": "^14.0.3",
+    "@angular/service-worker": "^14.0.3",
     "@ant-design/dark-theme": "^2.0.2",
     "@commitlint/cli": "^13.0.0",
     "@commitlint/config-angular": "^13.0.0",
@@ -117,7 +117,7 @@
     "marked": "^4.0.10",
     "minimist": "^1.2.5",
     "monaco-editor": "^0.33.0",
-    "ng-packagr": "^13.0.3",
+    "ng-packagr": "^14.0.2",
     "ngx-color": "^6.2.0",
     "node-prismjs": "^0.1.2",
     "parse5": "^6.0.1",
@@ -134,9 +134,9 @@
     "shelljs": "^0.8.4",
     "sitemap": "^4.1.1",
     "terser": "^5.5.1",
-    "ts-node": "~9.1.1",
+    "ts-node": "~10.8.1",
     "tslib": "^2.0.3",
-    "typescript": "~4.4.2",
+    "typescript": "~4.7.4",
     "yaml-front-matter": "^4.1.1",
     "zone.js": "~0.11.4"
   },
