@@ -1,5 +1,5 @@
 import { Component, OnInit, ViewChild, ViewEncapsulation } from '@angular/core';
-import { AbstractControl, FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
+import { AbstractControl, UntypedFormBuilder, UntypedFormControl, UntypedFormGroup, Validators } from '@angular/forms';
 
 import { NzMentionComponent } from 'ng-zorro-antd/mention';
 
@@ -35,14 +35,14 @@ import { NzMentionComponent } from 'ng-zorro-antd/mention';
 })
 export class NzDemoMentionFormComponent implements OnInit {
   suggestions = ['afc163', 'benjycui', 'yiminghe', 'RaoHai', '中文', 'にほんご'];
-  validateForm!: FormGroup;
+  validateForm!: UntypedFormGroup;
   @ViewChild('mentions', { static: true }) mentionChild!: NzMentionComponent;
 
   get mention(): AbstractControl {
     return this.validateForm.get('mention')!;
   }
 
-  constructor(private fb: FormBuilder) {}
+  constructor(private fb: UntypedFormBuilder) {}
 
   ngOnInit(): void {
     this.validateForm = this.fb.group({
@@ -50,7 +50,7 @@ export class NzDemoMentionFormComponent implements OnInit {
     });
   }
 
-  mentionValidator = (control: FormControl): { [s: string]: boolean } => {
+  mentionValidator = (control: UntypedFormControl): { [s: string]: boolean } => {
     if (!control.value) {
       return { required: true };
     } else if (this.mentionChild.getMentions().length < 2) {
