@@ -8,7 +8,7 @@ import { PlatformModule } from '@angular/cdk/platform';
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
-import { NzFormItemFeedbackIconComponent, NzFormPatchModule } from 'ng-zorro-antd/core/form';
+import { NzFormPatchModule } from 'ng-zorro-antd/core/form';
 import { NzOutletModule } from 'ng-zorro-antd/core/outlet';
 import { NzIconModule } from 'ng-zorro-antd/icon';
 
@@ -34,7 +34,6 @@ import { NzTextareaCountComponent } from './textarea-count.component';
     NzAutosizeDirective,
     NzInputGroupWhitSuffixOrPrefixDirective
   ],
-  entryComponents: [NzFormItemFeedbackIconComponent],
   imports: [BidiModule, CommonModule, NzIconModule, PlatformModule, NzOutletModule, NzFormPatchModule]
 })
 export class NzInputModule {}
